Use std::uintptr_t instead of std::uint32_t for peripheral addresses

+++ common/hal/clint/clint.hpp
@@ -21,7 +21,7 @@ namespace hal
          */
         constexpr
         explicit
-        clint(const std::uint32_t base_addr) :
+        clint(const std::uintptr_t base_addr) :
             m_base_addr{ base_addr }
         {
         }
@@ -64,7 +64,7 @@ namespace hal
         }
 
     private:
-        std::uint32_t m_base_addr = 0;
+        std::uintptr_t m_base_addr = 0;
     };
 
 }
