+++ common/hal/plic/plic.hpp
@@ -14,7 +14,7 @@ namespace hal
          */
         constexpr
         explicit
-        plic(const std::uint32_t base_addr) :
+        plic(const std::uintptr_t base_addr) :
             m_base_addr{ base_addr }
         {
         }
@@ -51,7 +51,7 @@ namespace hal
         }
 
     private:
-        std::uint32_t m_base_addr = 0;
+        std::uintptr_t m_base_addr = 0;
     };
 
 }
